+++ signaldata.h
@@ -23,13 +23,13 @@
 class SignalData
 {
 public:
-	SignalData(uint64_t samplerate);
+	SignalData(double samplerate);
 
 public:
-	uint64_t get_samplerate() const;
-	int64_t get_start_time() const;
+	double get_samplerate() const;
+	double get_start_time() const;
 
 protected:
-	const uint64_t _samplerate;
-	const int64_t _start_time;
+	const double _samplerate;
+	const double _start_time;
 };
