+++ sigview.h
@@ -55,8 +55,8 @@ private slots:
 private:
 	SigSession &_session;
 
-	uint64_t _scale;
-	int64_t _offset;
+	double _scale;
+	double _offset;
 };
 
 #endif // SIGVIEW_H
